Prestige Roshanara Bangalore - New Launch Apartments In Whitefield
Prestige Roshanara is an upcoming project in Whitefield, Bangalore developed by Prestige Groups. The developer is one of the best builders in the Whitefield area. The property offers 1/2/3 BHK apartments that are designed to offer best-in-class niceties to its purchasers. 

The value of the properties in Whitefield has seen a drastic increase in recent periods. Potential buyers have shown much interest in the residential properties of Whitefield exclusively, Prestige Roshanara Whitefield. 

The residential project is spread across wide acres of land area. The project is strategically positioned in the Whitefield which is in a significant upturn due to the development of residential spaces ranging from small affordable units to best properties. The project site is located in the eastern zone of Bangalore, this region is witnessing rapid development and therefore it has gained the interest of the top most builders to capitalize. The important factors that are responsible for the rapid growth of the residential market are its close immediacy to popular software companies, good connectivity, good socio-economic conditions, and improved infra conveniences. Apartments in Whitefield are good enough to buy because they house all the conveniences.

Project Highlights:-

✔ 1, 2, & 3 Bhk Apartments	
✔ Located At Whitefield, Bangalore.
✔ Apartment complex loaded with all amenities and facilities	
✔ Certified earthquake-resistant structure
✔ Wide Acres of Land	
✔ Excellent Connectivity
✔ Adequate provision for fire fighting on each floor	
✔ Luxury Clubhouse With Entertainment And Leisure Facilities
✔Amphitheater, Skating Rink, Exercise station, Badminton Court, Basket Ball Practice Court & Convenient Shopping

For More Details:
Visit Here: https://bit.ly/3C9epSs